Added game loop, IsEndState, and available moves.
Determining if the state is an end state involves seeing if the current
player has any moves available.  This uses the GetAvailableMovesForPiece
method, but stops after getting the first move.

Also added an IsJumped property to Piece, since, technically, pieces
should not be removed from the board until the player's turn has
completed.  And, added classes for three different types of moves:
sliding a piece, jumping with a piece, and ending the turn.